split backend progress tests
